pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyshearlab"
version = "0.0.1"
authors = [
  { name="Stefan Loock" },
]
description = "Shearlets in python"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "scipy",
    "pillow",
    "matplotlib",
    "pyfftw",
]

[project.optional-dependencies]
test = [
    "pytest",
]

[project.urls]
"Homepage" = "https://github.com/suxh19/pyshearlab"
"Bug Tracker" = "https://github.com/suxh19/pyshearlab/issues"

[tool.setuptools]
packages = ["pyshearlab"]

[dependency-groups]
dev = [
    "torch>=1.13.1",
]
